backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
email-validator
geoalchemy2
bcrypt
python-jose[cryptography]
